examples/gemini-ai-embeddings-demo/requirements.txt
google.generativeai
peewee
tidb-vector